requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
pillow==11.1.0
numpy==2.2.2
opencv-python-headless==4.10.0.84
tqdm==4.67.1
torch
torchvision
transformers
pytorch-lightning
segmentation-models-pytorch
albumentations
pyyaml